backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.7
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
google-auth==2.29.0
itsdangerous==2.2.0
pypdf==4.2.0
python-magic==0.4.27
google-cloud-storage==2.16.0